requirements-all.txt
# ============================================
# REQUISITOS COMPLETOS - Segunda Opinión Médica
# Incluye desarrollo y producción
# ============================================

# ===================
# CORE - PRODUCCIÓN
# ===================
Django==6.0
asgiref==3.11.0
sqlparse==0.5.4
tzdata==2025.3

# ===================
# FORMULARIOS Y UI
# ===================
django-crispy-forms>=2.3
crispy-bootstrap5>=2025.6
django-extensions>=3.2.3

# ===================
# PROCESAMIENTO DE MEDIOS
# ===================
Pillow>=11.0.0
pydicom>=2.4.0

# ===================
# HTTP Y API
# ===================
requests>=2.32.5
djangorestframework>=3.14.0

# ===================
# CIFRADO Y SEGURIDAD
# ===================
django-fernet-fields==0.6
django-fsm>=2.8.0
django-guardian>=2.5.0
django-auditlog>=2.1.0

# ===================
# ALMACENAMIENTO
# ===================
boto3>=1.30.0

# ===================
# TAREAS ASÍNCRONAS
# ===================
celery>=5.3.0
redis>=4.5.0

# ===================
# CONFIGURACIÓN
# ===================
python-dotenv>=1.0.0

# ===================
# WEBSOCKETS / DJANGO CHANNELS
# ===================
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# ===================
# PROCESAMIENTO DE TEXTO
# ===================
regex>=2023.0

# ============================================
# DESARROLLO Y TESTING
# ============================================

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Development tools
django-debug-toolbar==4.2.0
ipython==8.18.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Performance monitoring
django-silk==5.1.0

# Email testing
django-mail-panel==4.0.0
